validation 책임을 객체로 이전
